single device

+++ python/sgl_jax/srt/utils/weight_utils.py
@@ -1068,9 +1068,10 @@ class WeightLoader:
 
                 return _load_slice
 
-            lazy_array = jax.make_array_from_callback(shape, sharding, _make_load_slice()).astype(
-                target_dtype
-            )
+            lazy_array = jax.make_array_from_callback(
+                shape, sharding, _make_load_slice(), dtype=target_dtype
+            ).astype(target_dtype)
+
 
             lazy_arrays.append(lazy_array)
